#151258154 Restructure directory to follow Flask standards
