+++ app/models/shopping_item.py
