+++ app/models/shopping_list.py
