+++ app/templates/static/css/items.css
@@ -1,0 +1,3 @@
+body{
+    background-image: url("background.jpg");
+}
